requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.2
Django==4.2.1
django-cors-headers==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
